Working: RegExp would make this more powerful
Web scraper. Is currently in a file downloader config. To make this file more versatile you can use something called Regular Expressions to
search the parsed html file. Thy can also be used to help format the download directory or data structure. 

Currently this is set up to download roughly 1,000 white papers and text files that are on the hacking site. to limit this you can specify 
the types of files to download or but it in a counting loop to run it for a proof of concept without acquiring many files.